Value range

+++ hooklib/src/main/cpp/includes/cast.h
@@ -68,7 +68,7 @@ namespace SandHook {
                 return -1;
             }
 
-            for (int i = 0; i <= len; i += step) {
+            for (size_t i = 0; i <= len; i += step) {
                 T current_value = *reinterpret_cast<T *>((size_t) start + i);
                 if (value == current_value) {
                     return i;
